+++ examples/waypoint/lib/app/waypoint_bootstrap.dart
@@ -1,4 +1,3 @@
-import 'package:alphax/alphax.dart';
 import 'package:alphax_native/alphax_native.dart';
 import 'package:flutter/foundation.dart';